[OF-1471] test: Test Runner Shakedowns
A couple of changes to the test runner that emerged
during integration.

1. I hit the classic problem of calling a private header
in a public one.

2. The process descriptors were too generic,
and I actually accidentlyoutput the same text during
debugging, which caused pain due to
how we monitor stdout to flag futures.
