+++ MultiplayerTestRunner/include/ProcessDescriptors.h
@@ -24,50 +24,54 @@ namespace MultiplayerTestRunner::ProcessDescriptors
  * Process descriptors are emitted in stdout. The string will be emitted followed by a newline.
  * These are intended to be used by invoking processes to manage when they conduct their test assertions.
  * These may be disabled by setting the `--descriptors` flag in the CLI.
+ * It's important that you don't output these strings to the stream except in the correct places, and
+ * definitely not more than once per process. Be careful with debug logging.
  */
 
 /*
  * Emitted when the test has completed its setup and is ready for any controlling process to run tests
  * assertions. This is the main one you'll want to use.
  */
-constexpr const char* READY_FOR_ASSERTIONS_DESCRIPTOR = "READY_FOR_ASSERTIONS";
+constexpr const char* READY_FOR_ASSERTIONS_DESCRIPTOR = "READY_FOR_ASSERTIONS_DESCRIPTOR";
 
 /*
  *  Emitted when the test has logged in.
  */
-constexpr const char* LOGGED_IN_DESCRIPTOR = "LOGGED_IN";
+constexpr const char* LOGGED_IN_DESCRIPTOR = "LOGGED_IN_DESCRIPTOR";
 
 /*
  * Emitted when the test has logged out. Logout may not be emitted if process is terminated.
  */
-constexpr const char* LOGGED_OUT_DESCRIPTOR = "LOGGED_OUT";
+constexpr const char* LOGGED_OUT_DESCRIPTOR = "LOGGED_OUT_DESCRIPTOR";
 
 /*
  * Emitted when a new space is created. Not always emitted as existing space may be
  * specified for use via the `--space` CLI param.
  */
-constexpr const char* CREATED_SPACE_DESCRIPTOR = "CREATED_SPACE";
+constexpr const char* CREATED_SPACE_DESCRIPTOR = "CREATED_SPACE_DESCRIPTOR";
 
 /*
  * Emitted during cleanup when a space is deleted. Will not be emitted if the `--space` CLI param was used to
- * specify a custom space, as cleanup is not performed in that instance. Room cleanup may not be emitted if
+ * specify a custom space, as cleanup is not performed in that instance. Space cleanup may not be emitted if
  * process is terminated.
  */
-constexpr const char* DESTROYED_SPACE_DESCRIPTOR = "DESTROYED_SPACE";
+constexpr const char* DESTROYED_SPACE_DESCRIPTOR = "DESTROYED_SPACE_DESCRIPTOR";
 
 /*
  * Emitted when a logged in user joins a space
  */
-constexpr const char* JOINED_SPACE_DESCRIPTOR = "JOINED_SPACE";
+constexpr const char* JOINED_SPACE_DESCRIPTOR = "JOINED_SPACE_DESCRIPTOR";
 
 /*
  * Emitted when a user leaves a space
  */
-constexpr const char* EXIT_SPACE_DESCRIPTOR = "LEFT_SPACE";
+constexpr const char* EXIT_SPACE_DESCRIPTOR = "EXIT_SPACE_DESCRIPTOR";
 
-// This method exists just to encode the decision about there always being a newline.
+// This method exists to encode the decision about there always being a newline,
+// as well as to make sure we flush, which we need to do when using stdout as
+// an async channel of communication.
 inline void PrintProcessDescriptor(const char* Descriptor)
 {
-	std::cout << Descriptor << "\n";
+	std::cout << Descriptor << "\n" << std::flush;
 }
 } // namespace MultiplayerTestRunner::ProcessDescriptors
