+++ MultiplayerTestRunner/src/Utils.h
@@ -78,11 +78,6 @@ TestAccountCredentials LoadTestAccountCredentials();
  */
 void InitialiseCSPWithUserAgentInfo(const csp::common::String& EndpointRootURI);
 
-/*
- * Construct a new string from the input that is lower cased (via std::tolower)
- */
-std::string ToLowerCaseString(const std::string& input);
-
 constexpr char* DEFAULT_TEST_ENDPOINT			= "https://ogs-internal.magnopus-dev.cloud";
 constexpr int DEFAULT_TIMEOUT_IN_SECONDS		= 30;
 constexpr bool DEFAULT_EMIT_PROCESS_DESCRIPTORS = true;
